connect graphs
